backend/requirements.txt
crewai==0.14.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.8.1
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
pytest-asyncio>=0.23.3
httpx>=0.27.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.3
httpx>=0.27.0
